Small NFC to enable predicate printing in custom assembly

+++ iree/compiler/Dialect/VMLA/Conversion/StandardToVMLA/ConvertStandardToVMLA.cpp
@@ -102,7 +102,7 @@ struct CmpIOpConversion
     auto dst = VMLAConversionTarget::allocateOutputBuffer(
         srcOp.getLoc(), srcOp.getResult(), typeConverter, rewriter);
     auto newOp = rewriter.create<IREE::VMLA::CmpOp>(
-        srcOp.getLoc(), static_cast<uint32_t>(predicate), operands[0],
+        srcOp.getLoc(), predicate, operands[0],
         operands[1], dst, TypeAttr::get(inputType.getElementType()));
     if (forceUnsigned) {
       newOp.setAttr("force_unsigned", UnitAttr::get(rewriter.getContext()));
@@ -158,7 +158,7 @@ class CmpFOpConversion
     auto dst = VMLAConversionTarget::allocateOutputBuffer(
         srcOp.getLoc(), srcOp.getResult(), typeConverter, rewriter);
     auto newOp = rewriter.create<IREE::VMLA::CmpOp>(
-        srcOp.getLoc(), static_cast<uint32_t>(predicate), operands[0],
+        srcOp.getLoc(), predicate, operands[0],
         operands[1], dst, TypeAttr::get(inputType.getElementType()));
     rewriter.replaceOp(srcOp, newOp.dst());
     return success();
